requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==3.2.13
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mssql-django==1.0rc1
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.2
pyodbc==4.0.32
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
